requirements.txt
beautifulsoup4
pycryptodomex
requests
mpmath
